Apply internationalization to data in the database and update DTO and entity

+++ Jeju_Server/src/main/java/com/example/jeju_server/domain/dto/UserDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserDTO {
-    private Long id; // 유저의 고유 ID (optional, 사용자 생성 시에는 null일 수 있음)
+    private Integer id; // 유저의 고유 ID (optional, 사용자 생성 시에는 null일 수 있음)
     private String email; // 유저 이메일
     private String password; // 유저 비밀번호
     private String name; // 유저 이름
