+++ Jeju_Server/src/main/java/com/example/jeju_server/repository/AttractionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AttractionRepository extends JpaRepository<AttractionEntity, Long> {
+public interface AttractionRepository extends JpaRepository<AttractionEntity, Integer> {
 
 }
